Added notifications and changed app theme
Added notifications with badges. Handled in MainActivity. Monitors
changes in the database and adds badges to the bottom nav bar on the
appropriate changes to status. **NOTE** App theme was changed to
Theme.MaterialComponents derivative to accomodate BadgeDrawable object.
Colors were set to as close to the same as before.